Refactored Wind module to its own file
